opportunity style fixes
